requirements.txt
Pillow~=8.0.1
python-telegram-bot~=13.1
antlr4-python3-runtime~=4.9